+++ gfx/drivers_font/gl2_raster_font.c
@@ -85,7 +85,7 @@ static void gl2_raster_font_free(void *data,
 
 static bool gl2_raster_font_upload_atlas(gl2_raster_t *font)
 {
-   unsigned i, j;
+   int i, j;
    GLint  gl_internal          = GL_LUMINANCE_ALPHA;
    GLenum gl_format            = GL_LUMINANCE_ALPHA;
    size_t ncomponents          = 2;
@@ -178,7 +178,7 @@ error:
 }
 
 static int gl2_raster_font_get_message_width(void *data, const char *msg,
-      unsigned msg_len, float scale)
+      size_t msg_len, float scale)
 {
    const struct font_glyph* glyph_q = NULL;
    gl2_raster_t *font               = (gl2_raster_t*)data;
@@ -230,11 +230,11 @@ static void gl2_raster_font_draw_vertices(gl2_raster_t *font,
 }
 
 static void gl2_raster_font_render_line(gl2_t *gl,
-      gl2_raster_t *font, const char *msg, unsigned msg_len,
+      gl2_raster_t *font, const char *msg, size_t msg_len,
       GLfloat scale, const GLfloat color[4], GLfloat pos_x,
       GLfloat pos_y, unsigned text_align)
 {
-   unsigned i;
+   int i;
    struct video_coords coords;
    const struct font_glyph* glyph_q = NULL;
    GLfloat font_tex_coords[2 * 6 * MAX_MSG_LEN_CHUNK];
@@ -326,7 +326,7 @@ static void gl2_raster_font_render_message(
        !font->font_driver->get_line_metrics(font->font_data, &line_metrics))
    {
       gl2_raster_font_render_line(font->gl, font,
-            msg, (unsigned)strlen(msg), scale, color, pos_x,
+            msg, strlen(msg), scale, color, pos_x,
             pos_y, text_align);
       return;
    }
@@ -336,8 +336,8 @@ static void gl2_raster_font_render_message(
    for (;;)
    {
       const char *delim = strchr(msg, '\n');
-      unsigned msg_len  = delim
-         ? (unsigned)(delim - msg) : (unsigned)strlen(msg);
+      size_t msg_len    = delim
+         ? (delim - msg) : strlen(msg);
 
       /* Draw the line */
       gl2_raster_font_render_line(font->gl, font,
