+++ gfx/drivers_font/rsx_font.c
@@ -201,7 +201,7 @@ error:
 }
 
 static int rsx_font_get_message_width(void *data, const char *msg,
-      unsigned msg_len, float scale)
+      size_t msg_len, float scale)
 {
    const struct font_glyph* glyph_q = NULL;
    rsx_font_t *font   = (rsx_font_t*)data;
@@ -236,7 +236,7 @@ static int rsx_font_get_message_width(void *data, const char *msg,
 static void rsx_font_draw_vertices(rsx_font_t *font,
       const video_coords_t *coords)
 {
-   unsigned i;
+   int i;
    const float *vertex              = coords->vertex;
    const float *tex_coord           = coords->tex_coord;
    const float *color               = coords->color;
@@ -270,17 +270,17 @@ static void rsx_font_draw_vertices(rsx_font_t *font,
 }
 
 static void rsx_font_render_line(
-      rsx_font_t *font, const char *msg, unsigned msg_len,
+      rsx_font_t *font, const char *msg, size_t msg_len,
       float scale, const float color[4], float pos_x,
       float pos_y,unsigned text_align)
 {
-   unsigned i;
+   int i;
    struct video_coords coords;
    const struct font_glyph* glyph_q = NULL;
    float font_tex_coords[2 * 6 * MAX_MSG_LEN_CHUNK];
    float font_vertex[2 * 6 * MAX_MSG_LEN_CHUNK];
    float font_color[4 * 6 * MAX_MSG_LEN_CHUNK];
-   rsx_t      *rsx        = font->rsx;
+   rsx_t      *rsx      = font->rsx;
    const char* msg_end  = msg + msg_len;
    int x                = roundf(pos_x * rsx->vp.width);
    int y                = roundf(pos_y * rsx->vp.height);
@@ -376,8 +376,8 @@ static void rsx_font_render_message(
    for (;;)
    {
       const char *delim = strchr(msg, '\n');
-      unsigned msg_len  = delim
-         ? (unsigned)(delim - msg) : (unsigned)strlen(msg);
+      size_t msg_len    = delim
+         ? (delim - msg) : strlen(msg);
 
       /* Draw the line */
       rsx_font_render_line(font,
