+++ gfx/drivers_font/sixel_font.c
@@ -70,7 +70,7 @@ static void sixel_font_free(void *data, bool is_threaded)
 }
 
 static int sixel_font_get_message_width(void *data, const char *msg,
-      unsigned msg_len, float scale) { return 0; }
+      size_t msg_len, float scale) { return 0; }
 static const struct font_glyph *sixel_font_get_glyph(
       void *data, uint32_t code) { return NULL; }
