+++ gfx/drivers_font/vga_font.c
@@ -72,7 +72,7 @@ static void vga_font_render_free(void *data, bool is_threaded)
 }
 
 static int vga_font_get_message_width(void *data, const char *msg,
-      unsigned msg_len, float scale) { return 0; }
+      size_t msg_len, float scale) { return 0; }
 static const struct font_glyph *vga_font_get_glyph(
       void *data, uint32_t code) { return NULL; }
