+++ menu/drivers/ozone.c
@@ -5374,7 +5374,7 @@ border_iterate:
       {
          /* Note: This entry can never be selected, so ticker_x_offset
           * is irrelevant here (i.e. this text will never scroll) */
-         unsigned text_width = font_driver_get_message_width(ozone->fonts.entries_label.font, rich_label, (unsigned)strlen(rich_label), 1.0f);
+         unsigned text_width = font_driver_get_message_width(ozone->fonts.entries_label.font, rich_label, strlen(rich_label), 1.0f);
          x_offset = (video_info_width - (unsigned)
                ozone->dimensions_sidebar_width - entry_padding * 2) 
             / 2 - text_width / 2 - 60 * scale_factor;
@@ -6379,7 +6379,7 @@ static void ozone_draw_osk(ozone_handle_t *ozone,
                ? 0 
                : font_driver_get_message_width(
                      ozone->fonts.entries_label.font, msg,
-                     (unsigned)strlen(msg), 1.0f);
+                     strlen(msg), 1.0f);
             gfx_display_draw_quad(
                   p_disp,
                   userdata,
@@ -6491,7 +6491,7 @@ static void ozone_draw_messagebox(
       if (!string_is_empty(msg))
       {
          int width = font_driver_get_message_width(
-               ozone->fonts.footer.font, msg, (unsigned)strlen(msg), 1.0f);
+               ozone->fonts.footer.font, msg, strlen(msg), 1.0f);
 
          if (width > longest_width)
             longest_width = width;
@@ -8167,7 +8167,7 @@ static bool ozone_init_font(
    if (wideglyph_str)
    {
       int wideglyph_width =
-         font_driver_get_message_width(font_data->font, wideglyph_str, (unsigned)strlen(wideglyph_str), 1.0f);
+         font_driver_get_message_width(font_data->font, wideglyph_str, strlen(wideglyph_str), 1.0f);
       
       if (wideglyph_width > 0 && glyph_width > 0) 
          font_data->wideglyph_width = wideglyph_width * 100 / glyph_width;
@@ -8185,7 +8185,7 @@ static void ozone_cache_footer_label(ozone_handle_t *ozone,
 {
    const char *str = msg_hash_to_str(enum_idx);
    /* Determine pixel width */
-   unsigned length = (unsigned)strlen(str);
+   size_t length   = strlen(str);
 
    /* Assign string */
    label->str      = str;
